[rename] 파일 이름 및 폴더 구조 변경
Resolves: #34
