+++ product/products_data/test
@@ -1,0 +1,2 @@
+데이터 저장 경로 폴더 생성용 파일
+pull 받으신 후 이 test 파일을 삭제해 주세요!
